backend/requirements.txt
# Core backend
fastapi
uvicorn
python-dotenv
python-multipart

# LangChain + Embeddings + Retrieval
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
InstructorEmbedding
transformers
sentence-transformers  # still useful for fallback/default
rank_bm25

# DOM/XPath and HTML parsing
beautifulsoup4
lxml
